Fix PyCrypto dependency version in Windows
Fix the PyCrypto dependency to version 2.6 on Windows, which is what is
provided on voidspace.org.uk as a binary installer. Otherwise pip
tries to install the latest version from PyPI, even if the requirement
is already satisfied after running the PyCrypto binary installer.
Of course, installing the latest version from PyPI fails in the
absence of a C compiler.