Allow modules with custom CMake build systems to export include dirs.
Without this support the only way of doing conditional include directories is
to use targetDependencies, which requires splitting the header files into
separate modules.

Note that a need to use this functionality probably indicates you're doing
something wrong. It is, however, useful for porting software to yotta while
remaining compatible with an existing build system.