Don't send an invalid token to github.
In some circumstances we were sending "token None" as our authentication
information. Github behaviour seems to have recently changed so that even for
requests that do not require authentication, this became an error.

Additionally, make sure that errors are raised if the github tarball requests
fail in the future, and use the non-legacy urls for archives.