Add noninteractive mode
Activated by -n/--noninteractive. If authentication is required in
noninteractive mode, then a message like this will be displayed:

> error: login required (yotta is running in noninteractive mode)
> info: login URL: http://yottabuild.org/#login/abcde....

If multiple requests that yotta tried to perform require authentication, then
this might be displayed several times.

To make this possible, the way that some errors are handled in the
registry_access module has been adjusted, and the auth logic has been moved out
of github_access (where it didn't belong anyway) into a separate module.